requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
alembic==1.14.0
APScheduler==3.11.0
asyncio==3.4.3
attrs==24.2.0
bidict==0.23.1
blinker==1.9.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==44.0.0
Deprecated==1.2.15
dnspython==1.16.0
email-validator==1.3.1
Flask==3.1.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-SocketIO==5.4.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
frozenlist==1.5.0
gevent==24.11.1
gevent-websocket==0.10.1
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.8
MarkupSafe==3.0.2
multidict==6.1.0
numpy==1.24.3
packaging==24.2
pandas==2.2.3
propcache==0.2.1
psycopg2-binary==2.9.10
pycparser==2.22
PyGithub==2.5.0
PyJWT==2.10.1
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.10.1
python-socketio==5.11.4
pytz==2024.2
requests==2.32.3
schedule==1.2.2
simple-websocket==1.1.0
six==1.17.0
SQLAlchemy==2.0.36
ta==0.11.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
websocket-client==1.8.0
websockets==14.1
Werkzeug==3.1.0
wrapt==1.17.0
wsproto==1.2.0
WTForms==3.2.1
yarl==1.18.3
zope.event==5.0
zope.interface==7.2
psutil>=5.9.0


